# feat(desc): use concatenation instead of fmt.Sprintf to append strings